Fixed some file permissions

+++ Server_new_daemon/crc32.c
@@ -56,7 +56,7 @@
   /*                                                                        */
   /*  --------------------------------------------------------------------  */
 
-static unsigned long crc32_tab[] =
+static unsigned int crc32_tab[] =
 {
 	0x00000000L, 0x77073096L, 0xee0e612cL, 0x990951baL, 0x076dc419L,
 	0x706af48fL, 0xe963a535L, 0x9e6495a3L, 0x0edb8832L, 0x79dcb8a4L,
@@ -114,10 +114,10 @@ static unsigned long crc32_tab[] =
 
 /* Return a 32-bit CRC of the contents of the buffer. */
 
-unsigned long crc32(const unsigned char *s, unsigned int len)
+unsigned int crc32(const unsigned char *s, unsigned int len)
 {
 	unsigned int i;
-	unsigned long crc32val;
+	unsigned int crc32val;
 
 	crc32val = 0;
 	for (i = 0; i < len; i ++)
