+++ Server_new_daemon/crc32.h
@@ -31,6 +31,6 @@ Functions for computing 32-bit CRC.
 
 /* This computes a 32 bit CRC of the data in the buffer, and returns the
    CRC.  The polynomial used is 0xedb88320. */
-unsigned long crc32(const unsigned char *buf, unsigned int len);
+unsigned int crc32(const unsigned char *buf, unsigned int len);
 
 #endif /* CRC32_H */
